#1 fix: build the embedder without the cpu tag via a no-CGO stub engine
internal/embedder/embedder.go imported the llama-go bindings
unconditionally, so every package that links the embedder (cmd/hap,
internal/frontend, internal/embedder itself) pulled llama.cpp — and,
without the `cpu` tag, the GPU GGML backend — into its test binary. A
plain `go test ./...` on a CPU-only host therefore failed to LINK:

    /usr/bin/ld: cannot find -lggml-vulkan: No such file or directory
    /usr/bin/ld: cannot find -lvulkan: No such file or directory

(zgpu_linux.go in the bindings is //go:build !cpu and adds
`-lggml-vulkan -lvulkan`; the `cpu` tag selects the CPU backend instead.)

Split the engine on the `cpu` tag — the same tag every production/CI
build already passes, so build commands are unchanged:

- engine_cpu.go   (//go:build cpu)  — the real llama-linked Llama engine
  (moved verbatim out of embedder.go), which under `cpu` links only the
  CPU backend, never Vulkan.
- engine_stub.go  (//go:build !cpu) — a no-CGO Llama with the same method
  set: EmbedText reports unavailable, Dims()==0 (so the daemon's embed
  gate never routes to it), Degraded()==true. Links no native code.
- embedder.go stays tag-free with only the shared constants/helpers
  (Resolve*, PluginRoot, ErrDegraded) the Client and both engines use.

Result: a tag-free `go test ./...` compiles the stub and needs NO native
libraries; `go build -tags "vectors cpu"` links the real engine exactly
as before. No Vulkan library is required in any configuration.

Tests: the CGO/real-model cases move behind //go:build cpu
(engine_cpu_test.go, client_cpu_test.go); the engine-independent Client
supervision + protocol tests and TestResolveContextWindowFloor stay
untagged so they run in both builds; engine_stub_test.go adds coverage
for the no-CGO path (unavailable + Dims 0 + degraded).
